Applies Figaro config to heroku deployments
The Figaro config file is encrypted, and Travis config is updated
to decrypt the encrypted file as part of the build process.  The
Travis config also instructs Travis to skip cleanup after a build
so that the decrypted Figaro file will be included in the build
that is sent to heroku.

This commit also removes the secure environment variables from
Travis config, as they are already handled by Figaro.